--- src/components/DiskPageTitle/DiskPageTitle.scss
@@ -1,21 +1,0 @@
-@import '../../styles/mixins.scss';
-
-.ydb-disk-page-title {
-    display: flex;
-    flex-flow: row nowrap;
-    align-items: baseline;
-
-    text-wrap: nowrap;
-
-    @include header-2-typography();
-
-    &__prefix {
-        margin-right: 6px;
-
-        color: var(--g-color-text-secondary);
-    }
-
-    &__icon {
-        margin-right: 8px;
-    }
-}
