+++ src/components/EntityPageTitle/EntityPageTitle.scss
